pyproject.toml
[tool.poetry]
name = "camera-image"
version = "0.1.0"
description = ""
authors = ["Alexandr Vetrov <ernestkahnwald@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
djangorestframework = "^3.13.1"
ipdb = "^0.13.9"
opencv-python = "^4.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
